requirements.txt
openai
json
pandas
python-dotenv
streamlit
Pillow